+++ packages/web-ui/src/components/BaseRadioGroup/BaseRadioGroup.tsx
@@ -11,9 +11,9 @@ import { FieldsetLegend } from '../FieldsetLegend';
 import { Flex } from '../Flex';
 import { HelperText } from '../HelperText';
 
+import { mergeIds } from '../../helpers';
 import { useIds } from '../../hooks';
 import { PropsWithSx } from '../../types';
-import { mergeIds } from '../../utils';
 
 const componentName = 'BaseRadioGroup';
