+++ packages/web-ui/src/components/Text/Text.tsx
@@ -10,17 +10,11 @@ import { TextProps } from './Text.props';
 
 import { useBackground } from '../Box';
 
+import { DATA_ATTRIBUTES, DATA_ATTRIBUTE_SELECTORS, classSelector } from '../../helpers';
 import { styled } from '../../theme';
 import { fontWeights, fonts } from '../../tokens';
 import { PropsWithSx } from '../../types';
-import {
-  DATA_ATTRIBUTES,
-  DATA_ATTRIBUTE_SELECTORS,
-  classSelector,
-  mediaQueries,
-  pxToRem,
-  withGlobalPrefix,
-} from '../../utils';
+import { mediaQueries, pxToRem, withGlobalPrefix } from '../../utils';
 
 const componentName = 'Text';
 const componentClassName = withGlobalPrefix(componentName);
