+++ packages/web-ui/src/helpers/data-attributes.ts
@@ -2,7 +2,7 @@ function withDataPrefix(name: string) {
   return `data-${name}`;
 }
 
-/* a set of data-attributes used internally for styling */
+/* a set of data-attributes used for styling */
 export const DATA_ATTRIBUTES = {
   customColor: withDataPrefix('custom-color'),
   colorscheme: withDataPrefix('colorscheme'),
