+++ packages/web-ui/src/helpers/with-breakpoints.ts
@@ -1,6 +1,5 @@
-import { withGlobalPrefix } from './utils';
-
 import { Breakpoints, Responsive } from '../types';
+import { withGlobalPrefix } from '../utils';
 
 /** Translates a responsive object of boolean values to a responsive object of string values. To be used with `withBreakpoints`. */
 export const translateBooleanValues = (
