+++ packages/web-ui/src/index.ts
@@ -5,4 +5,6 @@ export * from './theme';
 export * from './tokens';
 export * from './types';
 export * from './hooks';
-export { mediaQueries, px, pxToRem, spacing, withGlobalPrefix, DATA_ATTRIBUTES } from './utils';
+export * from './utils';
+
+export { DATA_ATTRIBUTES } from './helpers';
