--- packages/web-ui/src/utils/color.ts
@@ -1,5 +1,0 @@
-import { type BackgroundColor, inverseBackgroundColors } from '../types';
-
-export const isInverseBackgroundColor = (backgroundColor: BackgroundColor): boolean => {
-  return (inverseBackgroundColors as ReadonlyArray<string>).includes(backgroundColor);
-};
