+++ packages/web-ui/src/utils/index.ts
@@ -1,9 +1,3 @@
-export * from './color';
-export * from './columns';
-export * from './css-selector-helpers';
-export * from './data-attributes';
 export * from './media-queries';
-export * from './merge-ids';
 export * from './typography';
 export * from './utils';
-export * from './with-breakpoints';
