+++ packages/web-ui/src/utils/typography.ts
@@ -2,8 +2,3 @@
 export const htmlFontSize = 16; // px
 /* converts a pixel font size to rems */
 export const pxToRem = (size: number) => `${size / htmlFontSize}rem`;
-
-export const isHeadingVariant = (variant: string): boolean => {
-  const headingVariants = ['displayHeading', 'h1', 'h2', 'h3', 'h4'];
-  return headingVariants.includes(variant);
-};
